# openapi.yaml
...
openapi: 3.1.0
info:
  title: TorrGate API
  version: 1.0.0
  description: High-performance TypeScript API gateway and proxy for torrent trackers, adhering to the canonical Jackett REST v2.0 path space.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: /api/v2.0/indexers
    description: Canonical Jackett REST v2.0 endpoint space
security:
  - ApiKeyAuth: []
  - ApiKeyQuery: []
  - BearerAuth: []
  - {}
paths:
...
  /{indexer}/categories:
    get:
      summary: Get indexer category mappings
      description: Returns the list of categories and their numeric IDs for a specific indexer.
      operationId: getCategories
      parameters:
        - name: indexer
          in: path
          required: true
          description: Tracker indexer name or identifier
          schema:
            type: string
            example: "rutor"
      responses:
        "200":
          description: List of category mappings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryMapping"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /{indexer}/download:
    get:
      summary: Download .torrent file
      description: Proxies and streams .torrent file downloads from trackers requiring cookies or proxy access.
      operationId: downloadTorrent
      parameters:
        - name: indexer
          in: path
          required: true
          description: Tracker indexer name or identifier
          schema:
            type: string
            example: "rutracker-ru"
        - name: url
          in: query
          required: true
          description: Relative or absolute download URL for the .torrent file
          schema:
            type: string
            example: "https://rutracker.org/forum/dl.php?t=100001"
      responses:
        "200":
          description: Torrent file binary content
          content:
            application/x-bittorrent:
              schema:
                type: string
                format: binary
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "502":
          $ref: "#/components/responses/BadGateway"
...
components:
  responses:
    BadGateway:
      description: Bad gateway error when upstream tracker fails
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Target download host is not an allowed mirror of the indexer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalServerError:
      description: Unexpected internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Requested resource was not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    TooManyRequests:
      description: Too many failed API key attempts from this client; retry after the one-minute window
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    CategoryMapping:
      type: object
      description: Mapping between category ID and its display name
      required:
        - id
        - name
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Movies"
    ErrorResponse:
      type: object
      description: Standard API error structure
      required:
        - error
        - message
        - statusCode
        - success
      properties:
        error:
          type: string
          example: "NotFound"
        message:
          type: string
          example: "Indexer 'unknown' not found"
        statusCode:
          type: integer
          example: 404
        success:
          type: boolean
          example: false
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
      description: Optional API key passed via HTTP header
    ApiKeyQuery:
      type: apiKey
      in: query
      name: apikey
      description: Optional API key passed via query parameter
    BearerAuth:
      type: http
      scheme: bearer
      description: Optional API key passed as a Bearer token